OkMD

Documentation for OkMD.

OkMD.expr_item

OkMD.get_changelog

OkMD.getheaderstr

OkMD.headlevel

OkMD.islevel

OkMD.islevelleq

OkMD.myflat

OkMD.plain_string

OkMD.read_section

OkMD.stripheaderstring

OkMD.targetrange

OkMD.targetsection

OkMD.expr_item — Constant

Expression that matches item bullet of asterisk.

OkMD.get_changelog — Method

get_changelog(projectfile, logfile, nlevel; kwargs...) get the target section of logfile that matches the version number in projectfile, where the version number is assumed to be in the header of nlevel.

OkMD.getheaderstr — Method

Given a vector Any, getheaderstr(v::Vector) returns a (nested) strings of plain string, code, text field of Markdown.___.text.

OkMD.headlevel — Method

headlevel(::Markdown.Header{n}) returns n.

OkMD.islevel — Method

Given a object mdc, islevel(mdc, n) returns true if mdc is the type of Markdown.Header{n}

OkMD.islevelleq — Method

Given a object mdc, islevelleq(mdc, n) returns true if mdc is the type of Markdown.Header{x} where x ≤ n. In brief, it recursively finds if it is a header of higher level (smaller n) until n==0 (false is returned).

OkMD.myflat — Method

Flatten nested vectors to a vector of strings.

OkMD.plain_string — Method

plain_string(md2::Markdown.MD) convert item bullet of asterisk (*) into -.

Example

using Markdown, OkMD
md2 = """
Changelog
- v0.1.0
"""

open("changelog_for_test.md", "w") do io
    write(io, OkMD.plain_string(md2))
end

OkMD.read_section — Method

read_section(fpath::String, nlevel::Int, exprh::Regex; with_header = true, kwargs...) read Markdown file fpath of Header of nlevel matching exprh. It takes the same kwargs as targetrange and targetsection.

OkMD.stripheaderstring — Method

stripheaderstring(mdhd::Markdown.Header) returns plain string concatenating the text or code field in Markdown container objects of format (i.e., Markdown.Code, Markdown.Link, Markdown.Italic, Markdown.Bold).

See also OkMD.getheaderstr, OkMD.myflat.

OkMD.targetrange — Method

Given a Vector, targetrange(mdcs::Vector, nlevel, exprh::Regex; which_one = only) find the target Markdown.Header{nlevel} object whose content matches exprh, returning a range which starts from this header until (but not include) the next header Markdown.Header{nlevelnext} where nlevelnext ≤ nlevel.

By default, if there is no or more than one header of nlevel matching exprh, error will be raised by which_one = only. You can assign which_one = last for example to get the last matched section.

Also see islevel, islevelleq and targetsection.

OkMD.targetsection — Method

Given a Markdown.MD object, targetsection(md1::Markdown.MD, nlevel, exprh) returns the section (which is a Vector) that starts with the Markdown.Header{nlevel} object whose content matches exprh and ends until the next header Markdown.Header{nlevelnext} where nlevelnext ≤ nlevel.

Also see targetrange.