OkMD

Documentation for OkMD.

OkMD.islevel

OkMD.islevelleq

OkMD.targetrange

OkMD.targetsection

OkMD.islevel — Method

Given a object mdc, islevel(mdc, n) returns true if mdc is the type of Markdown.Header{n}

OkMD.islevelleq — Method

Given a object mdc, islevelleq(mdc, n) returns true if mdc is the type of Markdown.Header{x} where x ≤ n. In brief, it recursively finds if it is a header of higher level (smaller n) until n==0 (false is returned).

OkMD.targetrange — Method

Given a Vector, targetrange(mdcs::Vector, nlevel, exprh::Regex) find the target Markdown.Header{nlevel} object whose content matches exprh, returning a range which starts from this header until (but not include) the next header Markdown.Header{nlevelnext} where nlevelnext ≤ nlevel.

Also see islevel, islevelleq and targetsection.

OkMD.targetsection — Method

Given a Markdown.MD object, targetsection(md1::Markdown.MD, nlevel, exprh) returns the section (which is a Vector) that starts with the Markdown.Header{nlevel} object whose content matches exprh and ends until the next header Markdown.Header{nlevelnext} where nlevelnext ≤ nlevel.

Also see targetrange.